requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
cryptography==44.0.0
frozenlist==1.5.0
idna==3.10
jwt==1.3.1
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
typing_extensions==4.12.2
yarl==1.18.3

python-dotenv~=1.0.1